story-refactor(dm): MISRA C++ narrowing and member-init, README
Explicit casts for integer narrowings (fetch_add/fetch_sub, size_t arr
counts, ternary size()), NSDMI for unser-member classes (pool_id_,
MergeCell, TokenBucketRateLimiter), Breaker init-list typed zeros, and a
0U loop index. No behavior change; 14/14 tests + examples green.

+++ examples/hsm_protocol_demo.cpp
@@ -180,8 +180,8 @@ int main()
     coact::EventPool<kBlk, kCap> pool;
     pool.init(storage, sizeof(storage), coact::make_critical_section(pal));
 
-    ProtocolAo ao(kStates, sizeof(kStates) / sizeof(kStates[0]),
-                  kTransitions, sizeof(kTransitions) / sizeof(kTransitions[0]),
+    ProtocolAo ao(kStates, static_cast<uint16_t>(sizeof(kStates) / sizeof(kStates[0])),
+                  kTransitions, static_cast<uint16_t>(sizeof(kTransitions) / sizeof(kTransitions[0])),
                   kDisconnected, /*max_depth=*/3U);
 
     coact::Event init_e;
